+++ spotify/playlist.py
@@ -1,10 +1,9 @@
 from __future__ import unicode_literals
 
-import collections
 import logging
 
 import spotify
-from spotify import ffi, lib, serialized, utils
+from spotify import compat, ffi, lib, serialized, utils
 
 
 __all__ = [
@@ -735,7 +734,7 @@ class PlaylistOfflineStatus(utils.IntEnum):
     pass
 
 
-class _Tracks(utils.Sequence, collections.MutableSequence):
+class _Tracks(utils.Sequence, compat.MutableSequence):
 
     def __init__(self, session, playlist):
         self._session = session
@@ -754,14 +753,14 @@ class _Tracks(utils.Sequence, collections.MutableSequence):
             sp_track=lib.sp_playlist_track(sp_playlist, key), add_ref=True)
 
     def __setitem__(self, key, value):
-        # Required by collections.MutableSequence
+        # Required by collections.abc.MutableSequence
 
         if not isinstance(key, (int, slice)):
             raise TypeError(
                 'list indices must be int or slice, not %s' %
                 key.__class__.__name__)
         if isinstance(key, slice):
-            if not isinstance(value, collections.Iterable):
+            if not isinstance(value, compat.Iterable):
                 raise TypeError('can only assign an iterable')
         if isinstance(key, int):
             if not 0 <= key < self.__len__():
@@ -776,7 +775,7 @@ class _Tracks(utils.Sequence, collections.MutableSequence):
         del self[key]
 
     def __delitem__(self, key):
-        # Required by collections.MutableSequence
+        # Required by collections.abc.MutableSequence
 
         if isinstance(key, slice):
             start, stop, step = key.indices(self.__len__())
@@ -793,7 +792,7 @@ class _Tracks(utils.Sequence, collections.MutableSequence):
         self._playlist.remove_tracks(key)
 
     def insert(self, index, value):
-        # Required by collections.MutableSequence
+        # Required by collections.abc.MutableSequence
 
         self[index:index] = [value]