+++ docs/changelog.rst
@@ -12,6 +12,10 @@ Bug fix release.
 - Add support for Python 3.6 and 3.7. No changes was required, but the test
   suite now runs on these versions too.
 
+- On Python 3, import :class:`Iterable`, :class:`MutableSequence`, and
+  :class:`Sequence` from :mod:`collections.abc` instead of :mod:`collections`.
+  This fixes a deprecation warning on Python 3.7 and prepares for Python 3.8.
+
 
 v2.0.5 (2015-09-22)
 ===================